+++ app/src/main/java/com/ncusoft/ncudocmsapp/repository/course/StudentCourseDao.java
@@ -135,12 +135,12 @@ public class StudentCourseDao implements TableInterface{
     }
 
     @Override
-    public long update(DatabaseHelper databaseHelper, ContentValues contentValues) {
+    public int update(DatabaseHelper databaseHelper, ContentValues contentValues) {
         return 0;
     }
 
     @Override
-    public long update(ContentValues contentValues) {
+    public int update(ContentValues contentValues) {
         return 0;
     }
