+++ app/src/main/java/com/ncusoft/ncudocmsapp/repository/login/UserDao.java
@@ -119,12 +119,12 @@ public class UserDao implements TableInterface{
     }
 
     @Override
-    public long update(DatabaseHelper databaseHelper, ContentValues contentValues) {
+    public int update(DatabaseHelper databaseHelper, ContentValues contentValues) {
         return 0;
     }
 
     @Override
-    public long update(ContentValues contentValues) {
+    public int update(ContentValues contentValues) {
         return 0;
     }
